CreateCard state removed
