--- src/state/createCard/index.js
@@ -1,5 +1,0 @@
-import Duck from "redux-ducks-ts";
-
-const createCardDuck = new Duck("createCard", "healthy");
-
-export default createCardDuck;
