--- src/state/createCard/sagas/index.js
@@ -1,3 +1,0 @@
-import createCardSagas from "./createCard.saga";
-
-export const sagas = [...createCardSagas];
