--- src/state/createCard/selectors.js
@@ -1,8 +1,0 @@
-import { createSelector } from "reselect";
-
-const reducer = ({ CreateCard }) => CreateCard;
-
-export const CreateCardSelector = createSelector(
-	[reducer],
-	(CreateCard) => CreateCard?.cardForm,
-);
