--- src/state/createCard/services/index.js
@@ -1,4 +1,0 @@
-import { fakeHttpModule, TYPES } from "common/http";
-
-export const postCard = ({ payload: { token, data } }) =>
-	fakeHttpModule.post("v1/cards", token, data, TYPES.json);
