--- src/state/createCard/types.js
@@ -1,5 +1,0 @@
-import createCardDuck from ".";
-
-export const CREATE_CARD_REQUEST = createCardDuck.defineType("CREATE_CARD_REQUEST");
-export const CREATE_CARD_SUCCESS = createCardDuck.defineType("CREATE_CARD_SUCCESS");
-export const CREATE_CARD_FAILURE = createCardDuck.defineType("CREATE_CARD_FAILURE");
