+++ src/state/rootSaga.js
@@ -2,21 +2,12 @@ import { fork, all } from "redux-saga/effects";
 // aca tienen que importar todas las sagas del state,
 // con un alias porque todas se exportan como "sagas" por default
 
-import { sagas as CreateCardSagas } from "state/createCard/sagas";
-
 import { sagas as AuthSagas } from "state/auth/sagas";
-
 import { sagas as SearchSagas } from "state/search/sagas";
 import { sagas as UserSagas } from "state/user/sagas";
 import { sagas as CardsSagas } from "state/cards/sagas";
 
-const allSagas = [
-	...AuthSagas,
-	...CreateCardSagas,
-	...SearchSagas,
-	...UserSagas,
-	...CardsSagas,
-];
+const allSagas = [...AuthSagas, ...SearchSagas, ...UserSagas, ...CardsSagas];
 
 export default function* rootSaga() {
 	yield all(allSagas.map((saga) => fork(saga)));
